requirements.txt
fastapi
uvicorn
pydantic
requests
dotenv
beautifulsoup4
lxml
cloudscraper
python-dotenv
openai
cloudinary
python-multipart
moviepy
ffmpeg-python
ffmpeg
apscheduler
gTTS 
pydub
apscheduler
google-search-results
serpapi
symspellpy
trafilatura
tiktoken
weaviate-client==4.14.0
ijson
boto3
more_itertools